Organise folder structure
Note that test files directly in dom/test/ are not found by mocha now
there are also test files present in its subdirectory; presumably due to
a bug in how the ** wildcard is interpreted in
packages/*/test/**/*.test.[jt]s; I have been unable to find out exactly
why or in which module this occurs.
