+++ packages/dom/test/range/cartesian.test.ts
@@ -18,4 +18,44 @@
  * under the License.
  */
 
-export {};
+import { assert } from 'chai';
+import { product } from '../../src/range/cartesian';
+
+async function* gen1() {
+  yield 1;
+  yield Promise.resolve(2);
+  yield 3;
+}
+
+async function* gen2() {
+  yield 4;
+}
+
+async function* gen3() {
+  yield 5;
+  yield 6;
+}
+
+describe('cartesian', () => {
+  describe('product', () => {
+    it('yields the cartesian product of the yielded items', async () => {
+      const cart = product(gen1(), gen2(), gen3());
+
+      const expected = [
+        [1, 4, 5],
+        [2, 4, 5],
+        [3, 4, 5],
+        [1, 4, 6],
+        [2, 4, 6],
+        [3, 4, 6],
+      ];
+
+      const result: number[][] = [];
+      for await (let value of cart) {
+        result.push(value);
+      }
+
+      assert.sameDeepMembers(expected, result, 'yields the expected items');
+    });
+  });
+});
